Rewrite docs and move context file
